separate chunk and chunk parser
