+++ data/tensorflow/recordio_reader.cc
@@ -32,8 +32,8 @@ Status RecordIOReader::ReadRecord(string* record) {
   // It is RecordIOReader who is responsible for handling multi-chunks 
   // in single recordio file.
   if (ReachEndOfChunk()) {
-    chunk_.Reset();
-    chunk_.Parse(input_stream_, &offset_);
+    chunk_.reset(new Chunk());
+    parser_.Parse(input_stream_, &offset_, chunk_);
     ++cur_chunknum_;
 
     // Check the offset for each chunk.
@@ -41,7 +41,7 @@ Status RecordIOReader::ReadRecord(string* record) {
   }
 
   // Fetch the next record.
-  Status s = chunk_.Next(record);
+  Status s = chunk_->Next(record);
   if (!s.ok()) {
     return s;
   }
